% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpspline.R
\name{mpspline_rmse1}
\alias{mpspline_rmse1}
\title{calculate RMSE}
\usage{
mpspline_rmse1(s = NULL, p = NULL, var_name = NULL)
}
\arguments{
\item{s}{site data frame}

\item{p}{estimated spline params for site}

\item{var_name}{target variable}
}
\value{
length-2 named numeric - RMSE and RMSE scaled against input data's
  interquartile range.
}
\description{
Calculates Root Mean Squared Error (RMSE) for estimates on a single site
}
\note{
Useful for comparing the results of varying parameter `lam`.
}
\keyword{internal}
