% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpspline.R
\name{mpspline_tmse1}
\alias{mpspline_tmse1}
\title{calculate RMSE}
\usage{
mpspline_tmse1(s = NULL, p = NULL, var_name = NULL)
}
\arguments{
\item{s}{site data frame}

\item{p}{estimated spline params for site}

\item{var_name}{target variable}
}
\value{
numeric, tmse
}
\description{
Calculates Root Mean Squared Error (RMSE) for estimates on a single site
}
\note{
Useful for comparing the results of varying parameter `lam`. RMSE_IQR
is the RMSE scaled to the IQR of the inputs.
}
\keyword{internal}
