% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpspline.R
\name{mpspline}
\alias{mpspline}
\title{Spline discrete soils data}
\usage{
mpspline(obj = NULL, var_name = NULL, lam = 0.1, d = c(0, 5, 15,
  30, 60, 100, 200), vlow = 0, vhigh = 1000)
}
\arguments{
\item{obj}{data frame, matrix or SoilProfileCollection containing measured
soil attributes arranged by depth.}

\item{var_name}{length-1 character denoting the column name in obj in which target
data is stored.}

\item{lam}{number; smoothing parameter for spline. Defaults to 0.1}

\item{d}{sequential integer vector; denotes the output depth ranges in cm.
Defaults to `c(0, 5, 15, 30, 60, 100, 200)` after the globalsoilmap.net
specification, giving output predictions over intervals 0-5cm, 5-15cm,
etc.}

\item{vlow}{numeric; constrains the minimum predicted value to a realistic number.
Defaults to 0.}

\item{vhigh}{numeric; constrains the maximum predicted value to a realistic number.
Defaults to 1000.}
}
\value{
list of five data elements for each site - input data, predicted
  values at each cm down the profile, predicted values over `d` intervals,
  and TMSE.
}
\description{
This function implements the mass-preserving spline method of Bishop et al
(1999) - http://dx.doi.org/10.1016/S0016-7061(99)00003-8] for interpolating
between measured soils data parameters down a profile.
}
\examples{
dat <- data.frame("SID" = c( 1,  1,  1,  1,   2,   2,   2,   2),
                   "UD" = c( 0, 20, 40, 60,   0,  15,  45,  80),
                   "LD" = c(10, 30, 50, 70,   5,  30,  60, 100),
                  "VAL" = c( 6,  4,  3, 10, 0.1, 0.9, 2.5,   6),
                   stringsAsFactors = FALSE)
mpspline(obj = dat, var_name = 'VAL')
}
