% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpspline.R
\name{mpspline_datchk}
\alias{mpspline_datchk}
\title{pre-spline data checks}
\usage{
mpspline_datchk(s = NULL, var_name = NULL)
}
\arguments{
\item{s}{data frame, input data for a single soil profile.}

\item{var_name}{character or integer vector denoting the column(s) in
\code{site} in which target data is stored. If not supplied, the fourth
column of the input object is assumed to contain the target data.}
}
\value{
If data passes checks it is returned unchanged. Sites with no data to
  spline and sites with overlapping input depth ranges return NA.
}
\description{
Runs a few data quality checks and makes some repairs where possible.
}
\keyword{internal}
