% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpspline.R
\name{mpspline_datchk}
\alias{mpspline_datchk}
\title{pre-spline data checks}
\usage{
mpspline_datchk(sites = NULL, var_name = NULL)
}
\arguments{
\item{sites}{list of appropriately formatted site data. Sites with no data to
spline and sites with overlapping depth ranges return NA.}

\item{var_name}{target variable}
}
\description{
Runs a few data quality checks and makes some repairs where possible.
}
\keyword{internal}
