% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpspline.R
\name{mpspline_rmse1}
\alias{mpspline_rmse1}
\title{calculate RMSE}
\usage{
mpspline_rmse1(s = NULL, p = NULL, var_name = NULL)
}
\arguments{
\item{s}{data.frame; data for one site}

\item{p}{list; estimated spline parameters for one site from
\code{\link[mpspline2:mpspline_est1]{mpspline_est1}}}

\item{var_name}{character or integer vector denoting the column(s) in
\code{site} in which target data is stored. If not supplied, the fourth
column of the input object is assumed to contain the target data.}
}
\value{
length-2 named numeric - RMSE and RMSE scaled against input data's
  interquartile range.
}
\description{
Calculates Root Mean Squared Error (RMSE) for estimates on a single site
}
\note{
Useful for comparing the results of varying parameter \code{lam}.
}
\keyword{internal}
