% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpspline.R
\name{mpspline_tmse1}
\alias{mpspline_tmse1}
\title{calculate TSME}
\usage{
mpspline_tmse1(s = NULL, p = NULL, var_name = NULL, s2 = NULL)
}
\arguments{
\item{s}{site data frame}

\item{p}{estimated spline params for site}

\item{var_name}{target variable}

\item{s2}{numeric, 5\% of the variance for the parent dataset}
}
\value{
numeric, tmse
}
\description{
Calculates Total Mean Squared Error (TMSE) for a single site
}
\keyword{internal}
